return to long

+++ src/main/java/org/olezha/MultiplicativeSegments.java
@@ -32,23 +32,25 @@ public class MultiplicativeSegments {
     private static final String INTERVAL_IS_FOUND_PREFIX = "Minimum interval length: ";
     private static final String FOUND_INTERVALS_MESSAGE = "Found intervals:";
 
-    static List<String> findShortestIntervals(BigInteger k, long n, List<BigInteger> nList) {
+    static List<String> findShortestIntervals(long k, long n, List<Long> nList) {
         long interval = 1;
         List<String> possibleIntervals = new LinkedList<>();
         for (; interval < n - 1; interval++) {
             long position = 0;
 
-            ListIterator<BigInteger> firstIterator = nList.listIterator();
-            ListIterator<BigInteger> secondIterator = nList.listIterator();
+            ListIterator<Long> firstIterator = nList.listIterator();
+            ListIterator<Long> secondIterator = nList.listIterator();
             for (long i = 1; i < interval; i++)
                 secondIterator.next();
             for (; secondIterator.hasNext();) {
                 position++;
-                BigInteger first = firstIterator.next();
-                BigInteger second = secondIterator.next();
+                Long first = firstIterator.next();
+                Long second = secondIterator.next();
 
                 // TODO: too slow check
-                if (first.multiply(second).remainder(k).equals(BigInteger.ZERO))
+//                if (first.multiply(second).remainder(k).equals(BigInteger.ZERO))
+//                if (((first % k) * (second % k)) % k == 0)
+                if (BigInteger.valueOf(first).multiply(BigInteger.valueOf(second)).remainder(BigInteger.valueOf(k)).equals(BigInteger.ZERO))
                     possibleIntervals.add("[" + position + ", " + (position + interval - 1) + "]");
             }
 
@@ -69,11 +71,11 @@ public class MultiplicativeSegments {
 
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        BigInteger k = BigInteger.valueOf(scan.nextLong());
+        long k = scan.nextLong();
         long n = scan.nextLong();
-        List<BigInteger> nList = new LinkedList<>();
+        List<Long> nList = new LinkedList<>();
         while (scan.hasNextLong())
-            nList.add(BigInteger.valueOf(scan.nextLong()));
+            nList.add(scan.nextLong());
         scan.close();
 
         for (String s : findShortestIntervals(k, n, nList))
